moving things to their places
