+++ src/helpers/balance/formatBalance.ts
