+++ src/helpers/index.ts
@@ -1,17 +1,17 @@
 // @index('./*.(tsx|ts)', f => `export * from '${f.path}'`)
 export * from './accountsAreEqual'
-export * from './checkIfBalanceSufficient'
-export * from './contractCall'
+export * from './balance/checkIfBalanceSufficient'
+export * from './balance/formatBalance'
+export * from './balance/getBalance'
+export * from './balance/getPSP22Balances'
+export * from './balance/transferBalance'
+export * from './contracts/contractCall'
+export * from './contracts/deployContract'
+export * from './contracts/getAbi'
+export * from './contracts/getAbiMessage'
 export * from './decodeOutput'
-export * from './deployContract'
-export * from './formatBalance'
-export * from './getAbi'
-export * from './getAbiMessage'
-export * from './getBalance'
 export * from './getExtrinsicErrorMessage'
 export * from './getGasLimit'
-export * from './getPSP22Balances'
 export * from './initPolkadotJs'
-export * from './transferBalance'
 export * from './unwrapResult'
 // @endindex
