+++ src/hooks/broker/purchaseInteract.ts
@@ -1,0 +1,33 @@
+import { useCallback, useState } from 'react';
+import { BrokerTransactionType } from '../../types/PurchaseInteract'; // Define this type as per your needs
+import { useSubstrateState } from '../substrate-lib';
+
+export const usePurchaseInteract = () => {
+  const [status, setStatus] = useState<string | null>(null);
+  const [param, setParam] = useState<string>('');
+
+  const { api } = useSubstrateState();
+
+  const handleInputChange = useCallback((value: string) => {
+    setParam(value);
+  }, []);
+
+  const submitTransaction = useCallback(() => {
+    if (!api.tx.broker || !api.tx.broker.purchase) {
+      return;
+    }
+
+    // Assuming the transaction requires these parameters
+    const transactionDetails: BrokerTransactionType = {
+      palletRpc: 'broker',
+      callable: 'purchase',
+      inputParams: [param],
+      paramFields: [{ name: 'param', type: 'TYPE', optional: false }]
+    };
+
+    // Here, you would implement the logic to handle the transaction
+    // setStatus would be used to update the status after the transaction
+  }, [api, param]);
+
+  return { status, param, handleInputChange, submitTransaction };
+};
