+++ src/hooks/index.ts
@@ -1,7 +1,7 @@
 // @index('./*.(tsx|ts)', f => `export * from '${f.path}'`)
-export * from './getConstants'
-export * from './useBalance'
-export * from './useContract'
-export * from './usePSP22Balances'
+export * from './balance/useBalance'
+export * from './balance/usePSP22Balances'
+export * from './broker/getConstants'
+export * from './contracts/useContract'
 export * from './useRegisteredContract'
 // @endindex
