+++ src/types/index.ts
@@ -1,8 +1,8 @@
 // @index('./*.ts', f => `export * from '${f.path}'`)
 export * from './ChainAsset'
-export * from './GetConstants'
 export * from './SubstrateChain'
 export * from './SubstrateDeployment'
 export * from './SubstrateWallet'
 export * from './UseInkathonProviderContext'
+export * from './broker/GetConstants'
 // @endindex
